Variable names changed and command interface methods definitions
The prompt and set_params methods definitions were changed in the Command interface class to support the use of a dictionary to store the arguments' values instead of a list. Some variables names were changed.

The load_command function as well as the function used to extract the arguments from a string, both previously defined in main.py, were changed to its own python file in the utils directory.